Edits for clarity.
